pyproject.toml
[project]
name = "tju-expense"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "SUPER", email = "hi@repus.me" }
]
dependencies = [
    "python-dotenv>=1.0.1",
    "pandas>=2.2.3",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/tju_expense"]
